Streamlit/requirements.txt
azure-monitor-opentelemetry
Requests==2.32.3
streamlit==1.37.1
